Add files via upload
This is my students project. It is a robotic arm controlled by Leap Motion sensor using the scripts written in Python. It is necessary that the Python version is 2.7. The controller is an Arduino Uno whose code is also uploaded. I used QtDesigner to create graphical user interface with very simple 2D visualization of robot's current position. To create the arm you need to laser cut the two .svg files in corresponding acrylic thickness. Actuators are 3 MG996R servo motors placed in base of the arm and 2 MG90S micro servo motors for the gripper and its orientation. All the joints have ball bearings for smooth motion.